+++ src/main/java/com/nisovin/shopkeepers/config/Settings.java
@@ -187,6 +187,7 @@ public class Settings extends Config {
 
 	public static boolean enableCitizenShops = true;
 	public static EntityType defaultCitizenNpcType = EntityType.PLAYER;
+	public static boolean saveCitizenNpcsInstantly = false;
 	public static boolean deleteInvalidCitizenShopkeepers = false;
 
 	public static boolean enableSignShops = true;